Addfacility: refactor - separate GUI from logic
