+++ src/group/project/recycling/POJOs/Facility.java
@@ -1,0 +1,13 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package group.project.recycling.POJOs;
+
+/**
+ *
+ * @author zoheb
+ */
+public class Facility {
+    
+}
